[Bug Fix] Fix session reseting on users
Fixed a bug where the session would be seemingly randomly reset.
29
Problem: As a part of the site we have the Web App Manifest. We don't really use it, but it enables our app as a Progressive Web App (more info about PWAs here ). However, when chromium makes a request for a manifest file, it doesn't send credientals, aka our cookie. Unforturnately, our web app took this as a new session (cause there was no cookie) and resets the session.
Solution: I've removed the web app manifest for now. There is a way to get it to use credentials but it wasn't working for me and I feel like there's been enough time lost to this problem that we can say we'll add it back in later cause I doubt anyone uses the PWA features of our site